refactor(core): move server routes around to clarify workspacing
