+++ packages/opencode/src/control-plane/workspace-context.ts
@@ -2,13 +2,13 @@ import { LocalContext } from "../util"
 import type { WorkspaceID } from "../control-plane/schema"
 
 export interface WorkspaceContext {
-  workspaceID: WorkspaceID
+  workspaceID: WorkspaceID | undefined
 }
 
 const context = LocalContext.create<WorkspaceContext>("instance")
 
 export const WorkspaceContext = {
-  async provide<R>(input: { workspaceID: WorkspaceID; fn: () => R }): Promise<R> {
+  async provide<R>(input: { workspaceID?: WorkspaceID; fn: () => R }): Promise<R> {
     return context.provide({ workspaceID: input.workspaceID }, () => input.fn())
   },
