+++ packages/opencode/src/server/routes/global.ts
