+++ packages/opencode/src/server/routes/instance/httpapi/config.ts
