+++ packages/opencode/src/server/routes/instance/httpapi/permission.ts
