+++ packages/opencode/src/server/routes/instance/httpapi/project.ts
