+++ packages/opencode/src/server/routes/instance/httpapi/provider.ts
