+++ packages/opencode/src/server/routes/instance/httpapi/question.ts
