+++ packages/opencode/src/server/routes/instance/httpapi/server.ts
