+++ packages/opencode/src/server/routes/instance/project.ts
@@ -1,13 +1,13 @@
 import { Hono } from "hono"
 import { describeRoute, validator } from "hono-openapi"
 import { resolver } from "hono-openapi"
-import { Instance } from "../../project/instance"
-import { Project } from "../../project"
+import { Instance } from "@/project/instance"
+import { Project } from "@/project"
 import z from "zod"
-import { ProjectID } from "../../project/schema"
-import { errors } from "../error"
-import { lazy } from "../../util/lazy"
-import { InstanceBootstrap } from "../../project/bootstrap"
+import { ProjectID } from "@/project/schema"
+import { errors } from "../../error"
+import { lazy } from "@/util/lazy"
+import { InstanceBootstrap } from "@/project/bootstrap"
 import { AppRuntime } from "@/effect/app-runtime"
 
 export const ProjectRoutes = lazy(() =>
