+++ packages/opencode/src/server/routes/instance/provider.ts
@@ -1,15 +1,15 @@
 import { Hono } from "hono"
 import { describeRoute, validator, resolver } from "hono-openapi"
 import z from "zod"
-import { Config } from "../../config"
-import { Provider } from "../../provider"
-import { ModelsDev } from "../../provider"
-import { ProviderAuth } from "../../provider"
-import { ProviderID } from "../../provider/schema"
-import { AppRuntime } from "../../effect/app-runtime"
+import { Config } from "@/config"
+import { Provider } from "@/provider"
+import { ModelsDev } from "@/provider"
+import { ProviderAuth } from "@/provider"
+import { ProviderID } from "@/provider/schema"
+import { AppRuntime } from "@/effect/app-runtime"
 import { mapValues } from "remeda"
-import { errors } from "../error"
-import { lazy } from "../../util/lazy"
+import { errors } from "../../error"
+import { lazy } from "@/util/lazy"
 import { Effect } from "effect"
 
 export const ProviderRoutes = lazy(() =>
