+++ packages/opencode/src/server/routes/instance/question.ts
@@ -2,11 +2,11 @@ import { Hono } from "hono"
 import { describeRoute, validator } from "hono-openapi"
 import { resolver } from "hono-openapi"
 import { QuestionID } from "@/question/schema"
-import { Question } from "../../question"
+import { Question } from "@/question"
 import { AppRuntime } from "@/effect/app-runtime"
 import z from "zod"
-import { errors } from "../error"
-import { lazy } from "../../util/lazy"
+import { errors } from "../../error"
+import { lazy } from "@/util/lazy"
 
 const Reply = z.object({
   answers: Question.Answer.zod
