+++ packages/opencode/src/server/routes/instance/session.ts
@@ -3,28 +3,28 @@ import { stream } from "hono/streaming"
 import { describeRoute, validator, resolver } from "hono-openapi"
 import { SessionID, MessageID, PartID } from "@/session/schema"
 import z from "zod"
-import { Session } from "../../session"
-import { MessageV2 } from "../../session/message-v2"
-import { SessionPrompt } from "../../session/prompt"
+import { Session } from "@/session"
+import { MessageV2 } from "@/session/message-v2"
+import { SessionPrompt } from "@/session/prompt"
 import { SessionRunState } from "@/session/run-state"
-import { SessionCompaction } from "../../session/compaction"
-import { SessionRevert } from "../../session/revert"
+import { SessionCompaction } from "@/session/compaction"
+import { SessionRevert } from "@/session/revert"
 import { SessionShare } from "@/share"
 import { SessionStatus } from "@/session/status"
 import { SessionSummary } from "@/session/summary"
-import { Todo } from "../../session/todo"
+import { Todo } from "@/session/todo"
 import { Effect } from "effect"
-import { AppRuntime } from "../../effect/app-runtime"
-import { Agent } from "../../agent/agent"
+import { AppRuntime } from "@/effect/app-runtime"
+import { Agent } from "@/agent/agent"
 import { Snapshot } from "@/snapshot"
-import { Command } from "../../command"
-import { Log } from "../../util"
+import { Command } from "@/command"
+import { Log } from "@/util"
 import { Permission } from "@/permission"
 import { PermissionID } from "@/permission/schema"
 import { ModelID, ProviderID } from "@/provider/schema"
-import { errors } from "../error"
-import { lazy } from "../../util/lazy"
-import { Bus } from "../../bus"
+import { errors } from "../../error"
+import { lazy } from "@/util/lazy"
+import { Bus } from "@/bus"
 import { NamedError } from "@opencode-ai/shared/util/error"
 import { jsonRequest } from "./trace"
