+++ packages/opencode/src/server/routes/ui.ts
